new PickupDetailView in new file
